:recycle: Move message types into their own subfolder
