+++ src/messages/internal/base_internal_event_message.ts
