+++ src/messages/internal/system_events/user_task_finished.ts
@@ -1,11 +1,12 @@
-import {UserTaskResult} from '../../../user_task/user_task_result';
+import {IIdentity} from '@essential-projects/iam_contracts';
 
-import {BasePublicEventMessage} from '../base_public_event_message';
+import {UserTaskResult} from '../../../data_models/user_task/user_task_result';
+import {BaseInternalEventMessage} from '../base_internal_event_message';
 
 /**
  * The message sent when a user task has been finished.
  */
-export class UserTaskFinishedMessage extends BasePublicEventMessage {
+export class UserTaskFinishedMessage extends BaseInternalEventMessage {
 
   /**
    * The result the user task was finished with.
@@ -18,8 +19,9 @@ export class UserTaskFinishedMessage extends BasePublicEventMessage {
               processInstanceId: string,
               flowNodeId: string,
               flowNodeInstanceId: string,
+              processInstanceOwner: IIdentity,
               currentToken: any) {
-    super(correlationId, processModelId, processInstanceId, flowNodeId, flowNodeInstanceId, currentToken);
+    super(correlationId, processModelId, processInstanceId, flowNodeId, flowNodeInstanceId, processInstanceOwner, currentToken);
 
     this.userTaskResult = userTaskResult;
   }
