+++ src/messages/internal/system_events/user_task_reached.ts
@@ -1,20 +1,19 @@
 import {IIdentity} from '@essential-projects/iam_contracts';
 
-import {UserTaskResult} from '../../../user_task/user_task_result';
+import {UserTaskResult} from '../../../data_models/user_task/user_task_result';
 import {BaseInternalEventMessage} from '../base_internal_event_message';
 
 /**
  * The message sent when a user task has been finished.
  */
-export class UserTaskFinishedMessage extends BaseInternalEventMessage {
+export class UserTaskReachedMessage extends BaseInternalEventMessage {
 
   /**
    * The result the user task was finished with.
    */
   public userTaskResult: UserTaskResult;
 
-  constructor(userTaskResult: UserTaskResult,
-              correlationId: string,
+  constructor(correlationId: string,
               processModelId: string,
               processInstanceId: string,
               flowNodeId: string,
@@ -22,7 +21,5 @@ export class UserTaskFinishedMessage extends BaseInternalEventMessage {
               processInstanceOwner: IIdentity,
               currentToken: any) {
     super(correlationId, processModelId, processInstanceId, flowNodeId, flowNodeInstanceId, processInstanceOwner, currentToken);
-
-    this.userTaskResult = userTaskResult;
   }
 }
