+++ src/messages/public/base_public_event_message.ts
